+++ src/lib/sentiment-parser.ts
@@ -1,10 +1,10 @@
-
-'use client';
+'use server';
 
 import type { AvatarState } from '@/types/dashboard';
 import { generateObject } from 'ai';
 import { google } from '@/lib/ai/groq';
 import { z } from 'zod';
+import { getEmotionFromTextByKeywords } from '@/lib/sentiment-parser.shared';
 
 const EmotionSchema = z.object({
     emotion: z.enum(['speaking_neutral', 'speaking_helpful', 'speaking_insightful', 'speaking_cautious'])
@@ -40,38 +40,4 @@ export async function getEmotionFromText(text: string): Promise<AvatarState> {
         // Fallback to keyword matching if the AI call fails
         return getEmotionFromTextByKeywords(text);
     }
-}
-
-
-const keywordMap: { keywords: string[]; state: AvatarState }[] = [
-  {
-    keywords: ['insight', 'interestingly', 'i have noticed', 'perhaps', 'consider', 'a pattern'],
-    state: 'speaking_insightful',
-  },
-  {
-    keywords: ['i can help', 'here is', 'of course', 'certainly', 'to do this', 'the result is'],
-    state: 'speaking_helpful',
-  },
-  {
-    keywords: ['warning', 'be careful', 'please note', 'however', 'be aware', 'cautious'],
-    state: 'speaking_cautious',
-  },
-];
-
-/**
- * (Fallback) Parses text using simple keywords to determine an appropriate emotional state for the avatar.
- * @param text The text to analyze.
- * @returns An AvatarState corresponding to the detected emotion.
- */
-export function getEmotionFromTextByKeywords(text: string): AvatarState {
-  const lowerText = text.toLowerCase();
-
-  for (const entry of keywordMap) {
-    if (entry.keywords.some((keyword) => lowerText.includes(keyword))) {
-      return entry.state;
-    }
-  }
-
-  // Default state if no specific keywords are found
-  return 'speaking_neutral';
 }
