+++ src/micro-apps/aegis-security/aegis-analysis-result.tsx
@@ -24,7 +24,6 @@ const containerVariants = {
     opacity: 1,
     transition: {
       staggerChildren: 0.1,
-      delayChildren: 0.2,
     },
   },
 };
